docs(contrast): generalize contrast audit, document install exception, fix doc drift
Phase 99 of docs/plans/clean-ui-1.2-seams (cross-cutting concerns surfaced
while specifying it, not blocking the seams themselves, but not worth
carrying silently into a two-package world).

- scripts/check-contrast.mjs was 487 lines entirely specific to clean-ui
  (hardcoded file paths, hardcoded theme list, an inline-duplicated semantic
  color scale) — the editor's syntax-highlighting requirement claims
  coverage by "the same contrast audit", which depended on this being
  possible. Split into scripts/contrast/color-math.mjs (generic WCAG math),
  scripts/contrast/clean-ui.mjs (clean-ui's own checks, now exporting run()
  instead of executing at import time), and a thin generic runner in
  check-contrast.mjs that aggregates any number of sources and names which
  one a failure came from. Verified byte-identical output for clean-ui's
  existing coverage (diffed against the pre-refactor script).
- CLAUDE.md: documented the global-install exception for satellite packages
  with a heavy dependency graph — an app.component() registration would put
  their whole dependency tree in every consumer's bundle and defeat
  route-level code splitting. Named-import + defineAsyncComponent example
  for both the satellite's own index.ts and a consuming app.
- Reconciled contributor docs with the current setup: the two claims this
  task named (docs site consuming a built library, release instructions
  predating the publish matrix) were already fixed incidentally in Phases 02
  and 03. Found one more in the broader sweep: CONTRIBUTING.md still said
  `pnpm test` runs all three vitest projects, which stopped being true when
  Phase 03 split the browser project into its own opt-in command. Fixed
  that, plus matching stale wording in CLAUDE.md's Build & Dev section, and
  actually ran every documented command (pnpm dev/build/test/check:contrast/
  verify:fixture) rather than only reading them.

Completes Phase 99 and all of docs/plans/clean-ui-1.2-seams (25/25 tasks,
4/4 phases).
